int/utils/sets: renames the package for its last symbol after removing unused ones.
